// projekt1-frontend/src/api/models/EditGameRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface EditGameRequest
 */
export interface EditGameRequest {
    /**
     * 
     * @type {number}
     * @memberof EditGameRequest
     */
    firstCompetitorScore?: number;
    /**
     * 
     * @type {number}
     * @memberof EditGameRequest
     */
    secondCompetitorScore?: number;
}

/**
 * Check if a given object implements the EditGameRequest interface.
 */
export function instanceOfEditGameRequest(value: object): boolean {
    let isInstance = true;

    return isInstance;
}

export function EditGameRequestFromJSON(json: any): EditGameRequest {
    return EditGameRequestFromJSONTyped(json, false);
}

export function EditGameRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): EditGameRequest {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'firstCompetitorScore': !exists(json, 'firstCompetitorScore') ? undefined : json['firstCompetitorScore'],
        'secondCompetitorScore': !exists(json, 'secondCompetitorScore') ? undefined : json['secondCompetitorScore'],
    };
}

export function EditGameRequestToJSON(value?: EditGameRequest | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'firstCompetitorScore': value.firstCompetitorScore,
        'secondCompetitorScore': value.secondCompetitorScore,
    };
}